+++ utils/schema.js
@@ -1,9 +1,0 @@
-'use strict';
-
-exports.getReferenceField = function (reference) {
-  return reference.split('.')[1];
-};
-
-exports.getReferenceCollectionName = function (reference) {
-  return reference.split('.')[0];
-};